removes nav-bad and consolidates them
