+++ src/app/admin/admin.component.html
@@ -1,4 +1,3 @@
-<quiz-app-nav-bar></quiz-app-nav-bar>
 <div class="background full-page">
   <mat-card class="matCard">
     <mat-card-header class="center">
