+++ src/app/learning-content-app/homepage/homepage.component.html
@@ -1,4 +1,4 @@
-<app-lc-nav-bar></app-lc-nav-bar>
+
 
 <!-- Hero image with login buttons -->
 <div class="hero">
