--- src/app/learning-content-app/lc-nav-bar/lc-nav-bar.component.scss
@@ -1,75 +1,0 @@
-.header{
-    background-color: #0e2340;
-    padding: 0px 10px 0px 10px;
-  
-  }
-  
-  .center {
-    margin-left: calc(50% - 150px);
-  }
-  
-  .main {
-    background-color: white !important;
-    width: 18%;
-    margin: 0 auto;
-    -moz-box-shadow:    3px 3px 1px 1px #EE7623;
-    -webkit-box-shadow: 3px 3px 1px 1px #EE7623;
-    box-shadow:         3px 3px 1px 1px #EE7623;
-  }
-  //#ee7623 #0e2340 #e6e7e8 #ffffff
-  
-  //fonts Montserrat light, semibold, regular
-  *{
-  margin: 0;
-  font-family: 'Montserrat', sans-serif;
-  font-weight: bold;
-  color: white;
-  }
-  .underline{text-decoration: underline;
-  
-  }
-  
-  .header{
-  height: 75px;
-  background-color: #0e2340;
-  display: flex;
-  }
-  
-  .headerLogo{
-  height: 60px;
-  margin-left: 30px;
-  margin-top: 15px;
-  }
-  
-  .navLink{margin-top: 25px;
-  margin-left: 25px;
-  text-align: right;
-  align-items: center
-  }
-  
-  .content{
-  display: flex;
-  }
-  .nameTitle{
-  margin-left: 20px;
-  font-size: 3vw;
-  }
-  
-  @media screen and (max-width: 800px) {
-  .headerLogo {
-    display: none;
-  }
-  }
-  
-  @media screen and (max-width: 600px) {
-  .nameTitle {
-    display: none;
-  }
-  .navLink {
-    margin: 0px;
-    padding: 5px;
-  }
-  .header {
-    height: 30px;
-  }
-  }
