+++ src/app/learning-content-app/learning-app.component.html
@@ -1,6 +1,6 @@
-<app-lc-nav-bar></app-lc-nav-bar>
+<!-- <app-lc-nav-bar></app-lc-nav-bar> -->
 <!-- <app-lc-nav-bar></app-lc-nav-bar>
 <learning-app-header></learning-app-header>
 <link href="https://fonts.googleapis.com/css?family=Roboto:300,400,500" rel="stylesheet">
 <learning-app-footer></learning-app-footer> -->
-<router-outlet></router-outlet>
+<!-- <router-outlet></router-outlet> -->
