+++ src/app/learning-content-app/learning-content-app.module.ts
@@ -26,7 +26,6 @@ import { LearningComponent } from './learning-app.component';
 import { SharedServicesModule } from '../shared/shared-service.module';
 import { CommonModule } from '@angular/common';
 import { AboutComponent } from './about/about.component';
-import { LcNavBarComponent } from './lc-nav-bar/lc-nav-bar.component';
 
 @NgModule({
   declarations: [
@@ -42,8 +41,7 @@ import { LcNavBarComponent } from './lc-nav-bar/lc-nav-bar.component';
     UploaderComponent,
     TestimonialComponent,
     TestimonialsComponent,
-    AboutComponent, 
-    LcNavBarComponent
+    AboutComponent
 
   ],
   imports: [
@@ -70,8 +68,7 @@ import { LcNavBarComponent } from './lc-nav-bar/lc-nav-bar.component';
     UploaderComponent,
     TestimonialComponent,
     TestimonialsComponent,
-    AboutComponent,
-    LcNavBarComponent
+    AboutComponent
   ]
   
 })
