+++ src/app/learning-content-app/public-content/public-content.component.html
@@ -1,4 +1,4 @@
-<app-lc-nav-bar></app-lc-nav-bar>
+
 <!-- test accordion -->
