+++ src/app/learning-content-app/testimonial/testimonial.component.html
@@ -1,4 +1,4 @@
-<app-lc-nav-bar></app-lc-nav-bar>
+
 <h1 class="header">Testimonial</h1>
 <div>
     <form [formGroup]="testimonialForm" (ngSubmit)="onSubmit()">
