+++ src/app/learning-content-app/testimonials/testimonials.component.html
@@ -1,4 +1,4 @@
-<app-lc-nav-bar></app-lc-nav-bar>
+
 <h2 class="header test">Testimonials</h2>
 
 <div class="list-group files-list">
