+++ src/app/learning-content-app/uploader/uploader.component.html
@@ -1,4 +1,4 @@
-<app-lc-nav-bar></app-lc-nav-bar>
+
 <main class="padding">
       <form class="create-form" [formGroup]="exampleForm" (ngSubmit)="onSubmit()">
         <div class="form-group">
