+++ src/app/learning-content-app/view-content/view-content.component.html
@@ -1,4 +1,4 @@
-<app-lc-nav-bar></app-lc-nav-bar>
+
 <ng-container *ngIf="items">
 <div class="container mt-3">
     <div class="title h3 mt-4">{{items.title}}</div>
