+++ src/app/quiz-app/dashboard/dashboard.component.html
@@ -1,4 +1,3 @@
-<quiz-app-nav-bar></quiz-app-nav-bar>
 <div class="background full-page">
   <div class="span"></div>
   <mat-card class="matCard" >
