+++ src/app/quiz-app/display-quiz/display-quiz.component.html
@@ -1,4 +1,3 @@
-<quiz-app-nav-bar></quiz-app-nav-bar>
 <form #questionForm="ngForm" class="background full-page">
   <div class="flex">
