+++ src/app/quiz-app/my-quizzes/my-quizzes.component.html
@@ -1,4 +1,3 @@
-<quiz-app-nav-bar></quiz-app-nav-bar>
 <mat-card class="matCard">
   <mat-card-header class="center">
     <mat-card-title class="fontSize">Your Quizzes</mat-card-title>
