--- src/app/quiz-app/nav-bar/nav-bar.component.html
@@ -1,18 +1,0 @@
-<div class="header">
-    <img class="headerLogo" src="./assets/images/quizWhipSemiBold300.png" alt="">
-    <div class="content" *ngIf="(user$ | async)">
-            <a class="navLink" [routerLink]="[ '/quiz/dashboard' ]" routerLinkActive="active">Dashboard</a>
-            <a class="navLink underline" (click)="takeQuiz()">Take Quiz</a>
-            <a class="navLink" [routerLink]="[ '/quiz/quiz-form' ]" routerLinkActive="active"
-                    *ngIf="(user$ | async).permission === 'admin' || (user$ | async).permission === 'teacher'">Create Quiz</a>
-            <a class="navLink" [routerLink]="[ '/admin' ]" routerLinkActive="active"
-                    *ngIf="(user$ | async).permission === 'admin'">Admin</a>
-                    <a class="navLink" [routerLink]="['/learning/homepage']">To Learning Content</a>
-                    <a class="navLink underline" (click)="signOut()">Sign Out</a>
-            
-            
-
-
-    </div>
-
-</div>
