+++ src/app/quiz-app/quiz-app.component.html
@@ -1,3 +1,2 @@
-<quiz-app-nav-bar></quiz-app-nav-bar>
 
 WORKING
