+++ src/app/quiz-app/quiz-form/quiz-form.component.html
@@ -1,4 +1,3 @@
-<quiz-app-nav-bar></quiz-app-nav-bar>
 <div class="flex background">
 <div class="full-page">
     <mat-card class="matCard createQuiz">
