+++ src/app/quiz-app/quiz-results/quiz-results.component.html
@@ -1,4 +1,3 @@
-<quiz-app-nav-bar></quiz-app-nav-bar>
 <h1>{{quiz.title}}</h1>
 
 <mat-card *ngFor="let question of quiz.questions">
